# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Amaris Service","description":"HTTP service for amaris test","version":"1.0"},"servers":[{"url":"http://localhost:8080"}],"paths":{"/friendlychains":{"post":{"tags":["amaris"],"summary":"FriendlyChains amaris","description":"Receives two words and returns if these are friendly","operationId":"amaris#FriendlyChains","requestBody":{"description":"The friendlychains request","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FriendlyChainsRequest"},"example":{"a":"to","b":"kyo"}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Friendlychainsresponse"},"example":{"result":"Las cadenas son amigas"}}}}}}},"/openapi.json":{"get":{"tags":["openapi"],"summary":"Download openapi3.json","operationId":"openapi#/openapi.json","responses":{"200":{"description":"File downloaded"}}}},"/pokemonname/{id}":{"get":{"tags":["amaris"],"summary":"PokemonName amaris","description":"Receives pokemon id and returns its name","operationId":"amaris#PokemonName","parameters":[{"name":"id","in":"path","description":"Pokemon id","required":true,"schema":{"type":"integer","description":"Pokemon id","example":2654504350665989948,"format":"int64"},"example":6314719275272383745}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Pokemonnameresponse"},"example":{"name":"Picachu"}}}}}}},"/sortnames":{"post":{"tags":["amaris"],"summary":"SortNames amaris","description":"Receives text string and returns two parameters: an array string and the mount of array","operationId":"amaris#SortNames","requestBody":{"description":"The sortnames request","required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SortNamesRequest"},"example":{"text":"Luis,Camilo,Andres,Laura"}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Sortnamesresponse"},"example":{"data":["Andres","Camilo","Laura","Luis"],"quantity":5}}}}}}}},"components":{"schemas":{"FriendlyChainsRequest":{"type":"object","properties":{"a":{"type":"string","description":"First word.","example":"to"},"b":{"type":"string","description":"Second word.","example":"kyo"}},"description":"Represents a friendlychains request.","example":{"a":"to","b":"kyo"},"required":["a","b"]},"Friendlychainsresponse":{"type":"object","properties":{"result":{"type":"string","description":"Friendly Chains","example":"Las cadenas son amigas"}},"example":{"result":"Las cadenas son amigas"}},"Pokemonnameresponse":{"type":"object","properties":{"name":{"type":"string","description":"Pokemon Name","example":"Picachu"}},"example":{"name":"Picachu"}},"SortNamesRequest":{"type":"object","properties":{"text":{"type":"string","description":"The text string.","example":"Luis,Camilo,Andres,Laura"}},"description":"Represents a sortnames request.","example":{"text":"Luis,Camilo,Andres,Laura"},"required":["text"]},"Sortnamesresponse":{"type":"object","properties":{"data":{"type":"array","items":{"type":"string","example":"Eveniet voluptatem nam dicta et veniam."},"description":"Names sort by first letter","example":["Andres","Camilo","Laura","Luis"]},"quantity":{"type":"integer","description":"Amount in Array","example":5,"format":"int64"}},"example":{"data":["Andres","Camilo","Laura","Luis"],"quantity":5}}}},"tags":[{"name":"openapi"},{"name":"amaris","description":"The amaris service provides endpoints to responde the amaris' test."}]}